+++ lib/activerecord_materialized_types.rb
@@ -5,6 +5,7 @@ module ActiveRecordMaterializedTypes
   DebounceInterval = T.type_alias { T.any(Integer, Float, ::ActiveSupport::Duration) }
   StalenessDuration = T.type_alias { T.any(Integer, ::ActiveSupport::Duration) }
   SourceDefinition = T.type_alias { T.any(String, Proc, T.nilable(T.proc.returns(String))) }
+  RefreshMode = T.type_alias { Symbol }
   RefreshCallbackName = T.type_alias { T.any(Symbol, Proc) }
   Connection = T.type_alias { ::ActiveRecord::ConnectionAdapters::AbstractAdapter }
   Timestamp = T.type_alias { T.any(::Time, ::ActiveSupport::TimeWithZone) }